refactor(client-cli): promote `certificate_client` module to directory
To split its code in more manageable part.
